Optional snat_ip parameter for dmz rules
Add a parameter to configure source IP translation in dmz rules. With
this set, connections to the backend host will have the source IP set to
this address instead of the originating IP.

The original rule in the POSTROUTING chain is not needed as this case is
already automatically covered by the connection tracking done by the
DNAT rule in the PREROUTING chain.